troupe: deliver signals before observing messages
When a process is both linked to as well as monitoring some other
process, we want to make sure any exception is propagated through the
link before a `Down` message can be observed by said process after which
it could, e.g., exit successfully, in case its monitor thread did not
(yet) observe the signal to inject.

This can be achieved by ensuring there are no pending signals at the
end of the `receiveWithOptions` implementation, before delivering any
matched message to the caller.

This patch includes a test-case which, before the fix, exposes the issue
(though this is a bug in concurrency, so it may succeed once in a
while). With the fix in place it should now succeed consistently.

Fixes: #25
See: https://github.com/NicolasT/troupe/issues/25